XMLBuilder: fix the falsy value bug
So, SimpleXML SUCKS. I've replaced it with DOMDocument. Most people
advise using DOMDOcument because SimpleXML has so many quirks and
limited / broken support for DOM modification.

I spent several hours futzing with it. Trying to add an
Element with no "value" or at least a falsy value. I was able to get it
to add the Element, but it would always mess up the rest of the
document. The doc would bw wrapped in a base element and the XML version
declaration would be at the top.

This now seems to work just fine, and the several are now fixed and
tested.